backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "bcrypt==5.0.0",
    "beanie>=2.0.1",
    "certifi==2025.11.12",
    "fastapi>=0.121.2",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "motor>=3.7.1",
    "passlib[bcrypt]==1.7.4",
    "pillow>=12.0.0",
    "pip==25.3",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-socketio>=5.14.3",
    "qrcode>=8.2",
    "uvicorn>=0.38.0",
]
